requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pandas>=2.0.0
numpy>=1.24.0
finnhub-python>=2.4.0
requests>=2.31.0
yfinance>=0.2.0
alpaca-trade-api>=3.0.0
